Particle: Port to c++
