+++ idlib/Makefile.Linux
@@ -38,9 +38,9 @@ all: $(IDLIB_TARGET)
 
 # Generic rule
 %.o: %.c
-	$(CXX) $(CXXFLAGS) -c $< -o $@
+	ccache $(CXX) $(CXXFLAGS) -c $< -o $@
 %.o: %.cpp
-	$(CXX) $(CXXFLAGS) -c $< -o $@
+	ccache $(CXX) $(CXXFLAGS) -c $< -o $@
 
 $(IDLIB_TARGET): $(OBJECTS)
 	ar rcsv $(IDLIB_TARGET) $(OBJECTS)
